[android] Fixed review notes
